@@ -10,7 +10,7 @@ description: A new Flutter project.
 version: 1.1.2
 
 environment:
-  sdk: ">=2.3.0-dev.0.0 <3.0.0"
+  sdk: ">=2.6.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -32,11 +32,11 @@ dependencies:
   flutter_sticky_header: ^0.4.0
   http:
   google_sign_in: ^4.0.1
-  intl: ^0.15.7
+  intl: ^0.16.0
   open_file: ^2.0.1
   package_info: ^0.4.0
   path_provider: ^0.5.0+1
-  rxdart: ^0.22.0
+  rxdart: ^0.23.1
   shared_preferences: ^0.5.2
   url_launcher: ^5.0.2
 
